DataMin2x/requirements.txt
google-generativeai>=0.3.0
asyncio
psutil>=5.8.0
aiofiles>=23.0.0
python-dateutil>=2.8.0
colorama>=0.4.6
tqdm>=4.64.0
requests>=2.28.0
beautifulsoup4>=4.11.0
nltk>=3.8
textstat>=0.7.0 